+++ src/test/java/vehiculos/Camioneta.java
@@ -7,7 +7,7 @@ public class Camioneta extends Vehiculo{
 	private boolean volco;
 	private static ArrayList<Camioneta> listado = new ArrayList<Camioneta>();
 	
-	public Camioneta(String placa, int puertas, String nombre, float precio, float peso, Fabricante fabricante) {
+	public Camioneta(String placa, int puertas, String nombre, int precio, int peso, Fabricante fabricante) {
 		super(placa, puertas, 90, nombre, precio, peso, "4X4", fabricante);
 		listado.add(this);
 	}
